% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{ClonalSimData}
\alias{ClonalSimData}
\title{Constructor for ClonalSimData}
\usage{
ClonalSimData(
  mutations = data.frame(),
  params = list(),
  clonal_structure = data.frame(),
  metadata = list()
)
}
\arguments{
\item{mutations}{data.frame with mutation data}

\item{params}{list with simulation parameters}

\item{clonal_structure}{data.frame with clone hierarchy}

\item{metadata}{list with additional metadata}
}
\value{
ClonalSimData object
}
\description{
Constructor for ClonalSimData
}
\examples{
# Typically created by simulateTumor(), but can be constructed manually
mutations <- data.frame(
  Mutation = "mut1",
  Chromosome = "chr1",
  Position = 1000000,
  Ref = "A",
  Alt = "T",
  True_VAF = 0.5,
  VAF = 0.48,
  Depth = 100,
  Alt_reads = 48,
  Clone = "Clone1",
  Type = "private"
)
params <- list(subclone_freqs = c(0.5))
sim_data <- ClonalSimData(mutations = mutations, params = params)

}
