% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getClonalStructure}
\alias{getClonalStructure}
\title{Get clonal structure}
\usage{
getClonalStructure(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
data.frame defining clone hierarchy
}
\description{
Get clonal structure
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
structure <- getClonalStructure(sim)
print(structure)

}
