% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getObservedVAF}
\alias{getObservedVAF}
\title{Get observed VAF values (with sequencing noise)}
\usage{
getObservedVAF(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
numeric vector of observed VAF values
}
\description{
Get observed VAF values (with sequencing noise)
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
obs_vaf <- getObservedVAF(sim)
summary(obs_vaf)

}
