% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getSimParams}
\alias{getSimParams}
\title{Get simulation parameters}
\usage{
getSimParams(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
list with simulation parameters
}
\description{
Get simulation parameters
}
\examples{
sim <- simulateTumor()
params <- getSimParams(sim)
params$subclone_freqs

}
