% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getTrueVAF}
\alias{getTrueVAF}
\title{Get true VAF values (biological, without sequencing noise)}
\usage{
getTrueVAF(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
numeric vector of true VAF values
}
\description{
Get true VAF values (biological, without sequencing noise)
}
\examples{
sim <- simulateTumor()
true_vaf <- getTrueVAF(sim)
summary(true_vaf)

}
