% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,ClonalSimData,missing-method}
\alias{plot,ClonalSimData,missing-method}
\title{Plot method for ClonalSimData}
\usage{
\S4method{plot}{ClonalSimData,missing}(x, y, type = "vaf_density", ...)
}
\arguments{
\item{x}{ClonalSimData object}

\item{y}{unused (for S4 compatibility)}

\item{type}{character indicating plot type: "vaf_density", "vaf_scatter",
"depth_histogram", "clone_matrix". Default is "vaf_density"}

\item{...}{additional arguments passed to plotting functions}
}
\value{
ggplot2 object
}
\description{
Plot method for ClonalSimData
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
plot(sim, type = "vaf_density")
plot(sim, type = "vaf_scatter")

}
