% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,ClonalSimData-method}
\alias{summary,ClonalSimData-method}
\title{Summary method for ClonalSimData}
\usage{
\S4method{summary}{ClonalSimData}(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
list with summary statistics
}
\description{
Summary method for ClonalSimData
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
summary(sim)

}
