% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{toGRanges}
\alias{toGRanges}
\title{Convert ClonalSimData to GRanges object}
\usage{
toGRanges(object, include_metadata = TRUE)
}
\arguments{
\item{object}{ClonalSimData object}

\item{include_metadata}{logical, include all metadata columns. Default: TRUE}
}
\value{
GRanges object with mutation information
}
\description{
Converts mutation data to a GenomicRanges::GRanges object,
  the standard Bioconductor data structure for genomic intervals.
}
\examples{
sim <- simulateTumor()
gr <- toGRanges(sim)
print(gr)

# Access metadata columns
GenomicRanges::mcols(gr)

# Subset by chromosome
gr_chr1 <- gr[GenomicRanges::seqnames(gr) == "chr1"]

}
