% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{toPyClone}
\alias{toPyClone}
\title{Convert to PyClone input format}
\usage{
toPyClone(object, file, sample_id = "sample1")
}
\arguments{
\item{object}{ClonalSimData object}

\item{file}{character, file path to write TSV}

\item{sample_id}{character, sample identifier. Default: "sample1"}
}
\value{
data.frame formatted for PyClone (invisibly)
}
\description{
Exports data in format suitable for PyClone clonal deconvolution
  analysis.
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))

\dontrun{
toPyClone(sim, file = "pyclone_input.tsv")
}

}
