% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata}
\usage{
getMetadata(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
list with metadata
}
\description{
Get metadata
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
metadata <- getMetadata(sim)
metadata$date

}
