% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{getMutations}
\alias{getMutations}
\title{Get mutations data frame}
\usage{
getMutations(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
data.frame with mutation information
}
\description{
Get mutations data frame
}
\examples{
sim <- simulateTumor()
mutations <- getMutations(sim)
head(mutations)

}
