% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{show,ClonalSimData-method}
\alias{show,ClonalSimData-method}
\title{Show method for ClonalSimData}
\usage{
\S4method{show}{ClonalSimData}(object)
}
\arguments{
\item{object}{ClonalSimData object}
}
\value{
NULL (prints to console)
}
\description{
Show method for ClonalSimData
}
\examples{
sim <- simulateTumor(subclone_freqs = c(0.3, 0.4, 0.3))
show(sim)

}
