% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioconductor.R
\name{toSciClone}
\alias{toSciClone}
\title{Convert to SciClone input format}
\usage{
toSciClone(object, file)
}
\arguments{
\item{object}{ClonalSimData object}

\item{file}{character, file path to write TSV}
}
\value{
data.frame formatted for SciClone (invisibly)
}
\description{
Exports data in format suitable for SciClone clonal analysis.
}
\examples{
sim <- simulateTumor()

\dontrun{
toSciClone(sim, file = "sciclone_input.tsv")
}

}
